valgrind: fix segmentation fault on JB
BZ: 79065

Valgrind was crashing systematically on JB.
This crash is caused by AT_IGNORE-ing AT_BASE.
This AT_IGNORE was needed to have breakpoints in shared libs
be handled properly (not very clear what is the problem
in the interaction between Valgrind GDBSERVER, AT_BASE and GDB).
Waiting to better understand all this, as a temporary bypass,
this patch ensures we do not ignore the AT_BASE on android.

This fix is from http://comments.gmane.org/gmane.comp.debugging.valgrind.devel/18903
and is ported to x86.
This fix is considered as a workaround until proper support provided by MR1
on JB

The possible consequence is that breakpoints might be inserted
by the Valgrind gdbserver at wrong addresses in shared lib.

Change-Id: I83e4d5722058aed608825939700a7e2a122c50eb
Orig-Change-Id: I9bb17d7907d8b9e663c7572a0649bf7df448a80a
Signed-off-by: Zhang Dongxing <dongxing.zhang@intel.com>
Signed-off-by: Christophe Guerard <christophe.guerard@intel.com>
Signed-off-by: Jin Wei <wei.a.jin@intel.com>
Reviewed-on: http://android.intel.com:8080/85707
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>