valgrind: fix segmentation fault on JB
BZ: 79065
Valgrind was crashing systematically on JB.
This crash is caused by AT_IGNORE-ing AT_BASE.
This AT_IGNORE was needed to have breakpoints in shared libs
be handled properly (not very clear what is the problem
in the interaction between Valgrind GDBSERVER, AT_BASE and GDB).
Waiting to better understand all this, as a temporary bypass,
this patch ensures we do not ignore the AT_BASE on android.
This fix is from http://comments.gmane.org/gmane.comp.debugging.valgrind.devel/18903
and is ported to x86.
This fix is considered as a workaround until proper support provided by MR1
on JB
The possible consequence is that breakpoints might be inserted
by the Valgrind gdbserver at wrong addresses in shared lib.

Change-Id: I1a2ee4f5a869cc46f126873728b9e33d7f669c87
Orig-Change-Id: Idbad82f578c0597a8d25f110ac30acb5467dd008
Category: aosp improvement
Domain: AOSP-Applications
Origin: internal
Upstream-Candidate: no
Signed-off-by: Zhang Dongxing <dongxing.zhang@intel.com>
Signed-off-by: Christophe Guerard <christophe.guerard@intel.com>
Signed-off-by: Jin Wei <wei.a.jin@intel.com>
Signed-off-by: Thomas Buhot <thomas.buhot@intel.com>